NGEN-114 Refactor regular string parsing to its class
